Eliminate 'possible loss of data' compiler warning in cordprnt (MS VC)
* cordprnt.c (CORD_vsprintf): Cast result of ec_len() to short if
long_arg<0.

+++ cord/cordprnt.c
@@ -232,7 +232,7 @@ int CORD_vsprintf(CORD * out, CORD format, va_list args)
                         } else {
                             short * pos_ptr;
                             pos_ptr = va_arg(args, short *);
-                            *pos_ptr = ec_len(result);
+                            *pos_ptr = (short)ec_len(result);
                         }
                         goto done;
                     case 'r':
